adding some basic tick-handling

+++ src/main/java/de/idrinth/gods_and_heroes/interfaces/Alignment.java
@@ -1,11 +1,13 @@
 package de.idrinth.gods_and_heroes.interfaces;
 
+import java.math.BigDecimal;
+
 public interface Alignment {
-    public int getIndividuality();
-    public int getCollectivism();
-    public int getCreation();
-    public int getDestruction();
-    public int getHonesty();
-    public int getFalsehood();
+    public BigDecimal getIndividuality();
+    public BigDecimal getCollectivism();
+    public BigDecimal getCreation();
+    public BigDecimal getDestruction();
+    public BigDecimal getHonesty();
+    public BigDecimal getFalsehood();
     public void merge(Alignment alignment);
 }
