+++ src/main/java/de/idrinth/gods_and_heroes/interfaces/Quest.java
@@ -1,10 +1,11 @@
 package de.idrinth.gods_and_heroes.interfaces;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface Quest {
-    public int getBelieveCost();
-    public int getRenownReward();
+    public BigDecimal getBelieveCost();
+    public BigDecimal getRenownReward();
     public List<Option> getOptions();
     public void chooseOption(Option option);
     public Option getChosenOption();
